Refactor batnav.ui package so it would be a bit more organised
