+++ src/batnav/ui/boards/PlayerBoard.java
@@ -1,9 +1,8 @@
-package batnav.ui;
+package batnav.ui.boards;
 
 import batnav.online.match.Bomb;
 import batnav.online.match.Match;
 import batnav.online.match.Ship;
-import batnav.utils.Colour;
 
 import java.awt.*;
